requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery~=5.1.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
Django~=3.2.8
django-appconf==1.0.5
django-celery-email==3.0.0
django-templated-mail==1.1.1
djangorestframework~=3.10.3
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==19.9.0
honcho==0.5.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.2
kombu==5.2.1
MarkupSafe==2.0.1
oauthlib==3.1.1
prompt-toolkit==3.0.22
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
rabbitmq==0.2.0
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5

views~=0.3
dj-config-url~=0.1.1
django-easy-pdf~=0.1.1
django-wkhtmltopdf~=3.3.0